fix: rename imports (#7)
